requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
apache-airflow==2.6.3
apache-airflow-providers-common-sql==1.5.2
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.4.2
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-postgres==5.5.1
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==40.0.2
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.2
executing==2.1.0
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
google-re2==1.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.6.0
numpy==2.1.1
ordered-set==4.1.0
packaging==21.3
pandas==2.2.2
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.2.0
prison==0.2.1
prompt_toolkit==3.0.47
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==26.2.0
referencing==0.29.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.4.2
rich-argparse==1.2.0
rpds-py==0.8.8
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.7.1
tzdata==2024.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
